requirements.txt
ccxt==4.2.67
numpy==1.24.3
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
